Test infra as well as a variety of tests to validate arrow ingestion / export.
The tests do not currently run on M1 macs, as the 9.0.0 arrow that is pulled
does not include a pre-built aarch64 binary (only includes intel).

This is probably not the right way to do a lot of this, but could potentially
be a way forward.

First compiles the jdbc driver and JNI bindings with no dependencies.
Then pulls down a variety of dependncies that are now required to compile the
  tests for the java system.
Then compiles the JDBC tests with these arrow dependencies
Then pulls down a variety of runtime dependencies for the tests
Writes out a shell file that can be used to execute the tests